Code cleanup on PersistentContainer
