+++ TweetNestKit/PersistentContainer.swift
@@ -20,55 +20,6 @@ public class PersistentContainer: NSPersistentCloudKitContainer {
         Session.containerApplicationSupportURL
     }
 
-    public static let managedObjectModel: NSManagedObjectModel = {
-        let managedObjectModel = NSManagedObjectModel(contentsOf: Bundle.tweetNestKit.url(forResource: Bundle.tweetNestKit.name!, withExtension: "momd")!)!
-
-        guard let accountEntity = managedObjectModel.entitiesByName["Account"] else {
-            fatalError("Account entity not found.")
-        }
-
-        guard let usersFetchedPropertyDescription = accountEntity.propertiesByName["users"] as? NSFetchedPropertyDescription else {
-            fatalError("users property not found in Account entity.")
-        }
-
-        usersFetchedPropertyDescription.fetchRequest?.sortDescriptors = [
-            NSSortDescriptor(keyPath: \User.creationDate, ascending: true),
-            NSSortDescriptor(keyPath: \User.modificationDate, ascending: true),
-        ]
-
-        guard let userEntity = managedObjectModel.entitiesByName["User"] else {
-            fatalError("User entity not found.")
-        }
-
-        guard let accountsFetchedPropertyDescription = userEntity.propertiesByName["accounts"] as? NSFetchedPropertyDescription else {
-            fatalError("accounts property not found in User entity.")
-        }
-
-        accountsFetchedPropertyDescription.fetchRequest?.sortDescriptors = [
-            NSSortDescriptor(keyPath: \Account.creationDate, ascending: true),
-        ]
-
-        return managedObjectModel
-    }()
-
-    class var defaultPersistentStoreURL: URL {
-        Self.defaultDirectoryURL().appendingPathComponent("TweetNestKit.sqlite")
-    }
-
-    static let defaultPersistentStoreConfiguration = "TweetNestKit"
-
-    class var accountsPersistentStoreURL: URL {
-        Self.defaultDirectoryURL().appendingPathComponent("Accounts.sqlite")
-    }
-
-    static let accountsPersistentStoreConfiguration = "Accounts"
-
-    class var dataAssetsPersistentStoreURL: URL {
-        Self.defaultDirectoryURL().appendingPathComponent("DataAssets.sqlite")
-    }
-
-    static let dataAssetsPersistentStoreConfiguration = "DataAssets"
-
     public override var viewContext: NSManagedObjectContext {
         let viewContext = super.viewContext
         viewContext.automaticallyMergesChangesFromParent = true
@@ -115,10 +66,11 @@ public class PersistentContainer: NSPersistentCloudKitContainer {
             self.usersSpotlightDelegate = UsersSpotlightDelegate(forStoreWith: tweetNestKitPersistentStoreDescription, coordinator: self.persistentStoreCoordinator)
             #endif
         } else {
-            persistentStoreDescriptions[0].configuration = nil
-            persistentStoreDescriptions[0].url = nil
-            persistentStoreDescriptions[0].type = NSInMemoryStoreType
-            persistentStoreDescriptions[0].cloudKitContainerOptions = nil
+            persistentStoreDescriptions.forEach {
+                $0.url = nil
+                $0.type = NSInMemoryStoreType
+                $0.cloudKitContainerOptions = nil
+            }
         }
     }
 
@@ -130,7 +82,6 @@ public class PersistentContainer: NSPersistentCloudKitContainer {
                 #if canImport(CoreSpotlight)
                 let dispatchGroup = DispatchGroup()
                 dispatchGroup.notify(queue: .global(qos: .default)) {
-
                     self.persistentStoreCoordinator.perform {
                         if let usersSpotlightDelegate = self.usersSpotlightDelegate {
                             usersSpotlightDelegate.startSpotlightIndexing()
